Renaming examples and combining network_validation, sample size  p value
