+++ src/eliater/frontdoor_backdoor/__init__.py
@@ -22,19 +22,15 @@ estimated using Pearl's frontdoor or backdoor criterion.
 from y0.examples import Example
 
 from .base import base_example
-from .multiple_mediators_single_confounder import multiple_mediators_single_confounder_example
-from .multiple_mediators_with_multiple_confounders import (
-    multiple_mediators_with_multiple_confounders_example,
-)
-from .multiple_mediators_with_multiple_confounders_nuisances import (
-    multiple_mediators_confounders_nuisance_vars_example,
-)
+from .example1 import multiple_mediators_single_confounder_example
+from .example2 import multiple_mediators_with_multiple_confounders_example
+from .example3 import multiple_mediators_confounders_nuisance_vars_example
 
 __all__ = [
     "base_example",
-    "multiple_mediators_single_confounder_example",
-    "multiple_mediators_with_multiple_confounders_example",
-    "multiple_mediators_confounders_nuisance_vars_example",
+    "example1",
+    "example2",
+    "example3",
 ]
 
 for x in __all__:
