+++ src/eliater/frontdoor_backdoor_discrete/__init__.py
@@ -2,12 +2,10 @@
 
 from y0.examples import Example
 
-from .single_mediator_with_multiple_confounders_nuisances_discrete import (
-    single_mediator_with_multiple_confounders_nuisances_discrete_example,
-)
+from .example1 import single_mediator_with_multiple_confounders_nuisances_discrete_example
 
 __all__ = [
-    "single_mediator_with_multiple_confounders_nuisances_discrete_example",
+    "example1",
 ]
 
 for x in __all__:
